requirements.txt
crewai
crewai[tools]
schedule
python-dotenv